SPIM-paper: changed appearance of absolute error in simulation data plot (points -> lines)
git-svn-id: https://www.dkfz.de/svn/B040/SIMFCS/trunk/diffusion4@787 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ fcsmeasurement.cpp
@@ -410,7 +410,7 @@ void FCSMeasurement::save() {
             int b=round(save_binning_time/corr_taumin);
             for (unsigned long long i=0; i<timesteps/b-1; i++) {
                 fprintf(f, "%15.10lf, %u\n", t, binned_timeseries[i]);
-                t=t+corr_taumin*b;
+                t=t+corr_taumin*(double)b;
             }
             fclose(f);
             std::string tsfn=fn;
